Revert "Big Patch - Working Mode 0 + Also Updated Up-sampling"
This reverts commit 732b4351f21458299a8bbfeab4b7b6c9154448c2.

+++ src/iofunc.h
@@ -25,6 +25,6 @@ void readBinData(const std::string, std::vector<float> &);
 
 void readRawData(const std::string, std::vector<uint8_t> &);
 
-void writeBinData(const std::string, const std::vector<double> &);
+void writeBinData(const std::string, const std::vector<float> &);
 
 #endif // DY4_IOFUNC_H
